Conflicts:
	libavcodec/h264_cabac.c
	libavcodec/h264_cavlc.c
	libavformat/matroskadec.c

+++ libavcodec/vc1dsp.h
@@ -29,9 +29,10 @@
 #define AVCODEC_VC1DSP_H
 
 #include "dsputil.h"
+#include "hpeldsp.h"
 #include "h264chroma.h"
 
-typedef void (*vc1op_pixels_func)(uint8_t *block/*align width (8 or 16)*/, const uint8_t *pixels/*align 1*/, int line_size, int h);
+typedef void (*vc1op_pixels_func)(uint8_t *block/*align width (8 or 16)*/, const uint8_t *pixels/*align 1*/, ptrdiff_t line_size, int h);
 
 typedef struct VC1DSPContext {
     /* vc1 functions */
